Remove Try-Except-Pass In Favor Of Testing Existence

+++ ekip/nationalparks/management/commands/field_trip.py
@@ -6,12 +6,12 @@ from nationalparks.management.commands import ficorcleaner as fc
 from nationalparks.models import FieldTripSite, BestVisitTime, YouthFacility
 
 #Begin by cleaning house!
-try:
+if FieldTripSite.objects.exists():
     FieldTripSite.objects.all().delete()
+if BestVisitTime.objects.exists():
     BestVisitTime.objects.all().delete()
+if YouthFacility.objects.exists():
     YouthFacility.objects.all().delete()
-except:
-    pass
     
 NAME_TO_ABBR = {r[1]: r[0] for r in FieldTripSite.AGENCY_CHOICES}
 NAME_TO_ABBR['Army Corps'] = 'USACE'
